Locking down mutations with authentication

+++ src/resolvers/mutation.js
@@ -1,6 +1,8 @@
 import bcrypt from 'bcryptjs';
 import jwt from 'jsonwebtoken';
 
+import getUserId from '../utils/getUserId';
+
 const Mutation = {
   async createUser(parent, args, {prisma}, info) {
 
@@ -50,32 +52,31 @@ const Mutation = {
       user,
       token: jwt.sign({userId: user.id}, 'thisisasecret')
     }
-
-
   },
-  async deleteUser(parent, args, {prisma}, info) {
+  async deleteUser(parent, args, {prisma, request}, info) {
 
-    const userExists = await prisma.exists.User({ id: args.id })
-    if(!userExists) {
-      throw new Error('User NOT Found')
-    }
+    const userId = getUserId(request)
 
     return prisma.mutation.deleteUser({
       where : {
-        id : args.id
+        id : userId
       }
     }, info)
   },
-  async updateUser(parent, args, {prisma}, info)  {
+  async updateUser(parent, args, {prisma, request}, info)  {
+
+    const userId = getUserId(request)
 
     return prisma.mutation.updateUser({
       where: {
-        id : args.id
+        id : userId
       },
       data: args.data
     }, info)
   },
-  createPost(parent, args, {prisma}, info) {
+  createPost(parent, args, {prisma, request}, info) {
+
+    const userId = getUserId(request)
 
     return prisma.mutation.createPost({
        data: {
@@ -84,13 +85,25 @@ const Mutation = {
          published: args.data.published,
          author: {
            connect: {
-             id: args.data.author
+             id: userId
            }
          }
        } 
       }, info)
   },
-  deletePost(parent, args, {prisma}, info) {
+  async deletePost(parent, args, {prisma, request}, info) {
+
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if(!postExists) {
+      throw new Error('Unable to delete post')
+    }
 
     return prisma.mutation.deletePost({
       where : {
@@ -98,7 +111,18 @@ const Mutation = {
       }
     }, info)
   },
-  updatePost(parent, args, {prisma}, info) {
+  async updatePost(parent, args, {prisma, request}, info) {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if(!postExists) {
+      throw new Error('Unable to update post')
+    }
 
     return prisma.mutation.updatePost({
       where: {
@@ -107,13 +131,15 @@ const Mutation = {
       data: args.data
     }, info)
   },
-  createComment(parent, args, {prisma}, info) {
+  createComment(parent, args, {prisma, request}, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         },
         post: {
@@ -124,7 +150,19 @@ const Mutation = {
       } 
      }, info)
   },
-  updateComment(parent, args, {prisma}, info) {
+  async updateComment(parent, args, {prisma, request}, info) {
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if(!commentExists) {
+      throw new Error('Unable to update comment')
+    }
+
     return prisma.mutation.updateComment({
       where: {
         id : args.id
@@ -132,7 +170,19 @@ const Mutation = {
       data: args.data
     }, info)
   },
-  deleteComment(parent, args, {prisma}, info) {
+  async deleteComment(parent, args, {prisma, request}, info) {
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if(!commentExists) {
+      throw new Error('Unable to delete comment')
+    }
+
     return prisma.mutation.deleteComment({
       where : {
         id : args.id
